Add further documentation and inline OutilineTreeWrapper into OutlineTree
